Optimize client to use less data
Upon connecting, no longer retrieve MAXLINES for every buffer, instead,
get the list of lines the first time you visit a buffer.  Additionally,
don't subscribe to the nicklist changes unless the buffer is open.

BufferLine - Now has a unique pointer, so we can check for duplicate
lines when adding them.  This prevents a request for the same line
resulting in it appearing twice.

RelayService - Add subscribeBuffer/unsubscribeBuffer functions to get
the list of lines and the nicklist on demand