Fix order of notification/increasing unread count
Observers should be notified after the number of unread messages has
been updated, so they can reset it if they'd like. Prior to this change,
if you were in a buffer, a line was added, and you went back to the
buffer list, you would see a 1 for unread messages.