Fix more null pointer exceptions
If there is no network available, then the InetAddress.getByName will
fail, leaving server = null.  This then causes a NullPointerException
when the Socket goes to use it. Modified how the server variable is
handled in order to fix this.

Also modified RelayConnection so that a handler can only be added once
for a specific message.  This should prevent the bufferManager receiving
multiple callbacks once the connection is restored.