Handle upgrades from weechat
The method of dealing with an Upgrade is to just disconnect, wait a few
moments, and then reconnect.  Should work for most situations.

This fixed a few bugs in LineHandler(where the buffer might not exist
anymore).

Also modified RelayConnection to allow message ID's that have no object
attached to them(in this case, the _upgrade, and _upgrade_ended
messages.

Fixes #14.