Removing menu-button-context and adds ActionBar actions for Menu and Hotlist
Following Google's advise that menu button navigation should be available in ActionBar instead for tablet users.
Started on Hotlist to be available from ActionBar. This works almost as it should, but need to force updating connectionStatus title in onPrepareOptionsMenu.
(button works, but wrong label, doesn't get updated)
Swapped out 'native' ActionBar with ActionBarSherlock to be compatible with pre-3.0