Make weechat-relay a normal jar instead of apklib
There is no reason for it to be an apklib, and it is causing problems in
eclipse by being one.  So now it is back to just being a normal jar.

Removed some android specific stuff from the library.  This library is
intended for general use, so android specific things will not be
accepted into it.