Bugfixes/Cleanup
Remove unused imports in a few places.
Prevent updateMenuContext from crashing if called without a menu
Keep track of fragment attached state in BufferListFragment
Remove connectionhandler when we unbind from the service in
BufferListFragment
Call onBuffersChanged during onConnect() in BufferListFragment, to fix
list not being rendered.
Notify that hotlist has changed when an item is removed from it
Add a few typechecks to ensure we don't do something bad