Update scrolling behaviour in buffer view
Lines now fill in from the bottom
If you are scrolled back reading, your view will not be reset to the
bottom when a new message is received.  This is the "transcript mode" of
a listview.

I did notice that my android 2.3 emulator does not play nicely with this
when sending messages(but when receiving it works fine).  Might just be
an emulator bug.

Fixes #44.