Fix crash if you return to the app
Sometimes when you return to the app, if the background service has
died(or crashed), when it restarts bufferManager will be empty, and a
NPE will occur when the app tries to show the buffer.  Check that we are
connected first before trying to retrieve the buffer.

This was fixed based on a crash report in the google play developer
console, so not all details regarding it are present.