Add preference to reduce network traffic
This preference will only work properly on weechat 0.4.1-dev compiled
after 2/3/2013 due to a bug in weechat core. If enabled on versions
prior to this buffers will not be updated when open.

Due to a secondary bug, notification is given for buffers opened when at
least one buffer is open in the client.  If no buffers are open, then
notifications for new buffers will not be received.  Additionaly,
_upgrade messages from the server will be missed as well.

Added a preference_long_summar.xml to res/layouts to allow for
preference items with more than a 2 line summary.

The upstream bug: http://savannah.nongnu.org/bugs/?38215
The commit that fixed it:
http://git.savannah.gnu.org/gitweb/?p=weechat.git;a=commitdiff;h=7a1f9a1d7f3e52ddc9174ef88f491b6d9446f7b7

Fixes #61