Relay Connection:
big changes. so functionally, its almost the same. The biggest thing is that i made this class more like a library. no external dependencies, and it only support weechat relay protocol, plain text and ssl connection. again i tried to remove as many variable as possible to take up the least amount of ram. moved the vars as close as possible to where they are used.