Fix wrong away-nick color caused by cache
This is a fixup for 7e89bb5. Android heavy cache makes the adapter
"leak" the away color to nicknames that are not away as the user
scrolls within the list (this recycles holders w/ their color, hence
the "leak").

To fix that, we cache the default color on first load and then apply it
on the textview for both away/online statuses.