partially revert an earlier commit—this wasn't the reason of the glitch
note to self: the “glitch” is caused when a buffer receives some messa-
ges but doesn't run runs its lines.onLinesListed(), and at the same time
user reads the buffer in WeeChat and WA receives an updated lastSeenLine

now, when opening a buffer in WA, because onLinesListed() wasn't run,
the buffer has no read marker. it is calculated and received and displa-
yed when onLinesListed() is called. this is mostly a non-issue