lose dependency on JUnit—implement our own AssertJ-style Assert class
the problem with JUnit is that it's a tad too big for the task; also,
LeakCanary doesn't want to work with junit on classpath. alternatively,
we could use Truth, but it depends on JUnit as well; or AssertJ, but it
adds over 7k methods, which bring the total number to over the 64k dex
limit.