don't track “others” (lines other than highlights and regular messages)
we aren't displaying this number, and it's only effectively used to
invalidate current read statistics when the numbers don't match those in
weechat's hotlist. the problem is that some messages, such as nick
changes, don't reliably make it to hotlist. so we can end up with more
“other” messages than weechat reports, which can lead to erroneous
invalidation.

since the situation where such invalidation can be useful is very rare,
namely, when the buffer loses its read status twice, or reuses the last
seen line pointer, the number of highlights and unreads should be enough
for a valid outcome