in BufferFragment, change onStart/onStop to onResume/onStop
when opening a buffer from notifications, onNewIntent is called after
onStart on Android 10. as by then the buffer fragments are already in
the ready state, this causes the current buffer to be marked as read
when switching to the one from the notification. it will clear the
notification for the current buffer as well.

alternative solution for this would be marking buffers as read after
a few seconds of the buffer having focus