keep done uploads (but not failed) when calculating progress ratio
before this, total ratio was calculated only for running uploads, which
resulted in the progress ring jumping forward and back. this changes it
to a less surprising behavior... unless an upload fails.

if an upload fails, then we can't use its progress as it will prevent
the total ratio from reaching 100%. we should then either remove it from
the pool or consider it done (uploaded to completion).

in the former case, the total ratio can go backwards. in the latter, it
can jump forward. for no particular reason i prefer the former