fix a crash in onEvent(ExceptionEvent) when creating dialog fragments
as of androidx.lifecycle version 2.3.0, LifecycleRegistry now verifies
that its methods are called on main thread. this lead to a crash when
trying to present a dialogue about the error.

the method onEvent can't simply be moved to the main thread as it calls
SSLHandler.checkHostnameAndValidity that does I/O.