on supported api, use getInsets() instead of systemWindowInsetTop() etc
on api 30, methods systemWindowInsetTop (now deprecated)
use a slightly different logic to calculate results.
it is apparent when you:

  * open a buffer
  * tap on input to open keyboard
  * turn off device via button
  * turn on device via button

this makes the input bar being slightly clipped by the keyboard.
this is because WindowInsets.Type.ime() is not used under the hood
for some reason.

using the new api we can specify WindowInsets.Type.ime() directly