Certificate dialog: use original exception; show actual trust chain, if any
When SSL connection fails, we get an exception. It doesn't include the
certificate chain. Before, the method that retrieves it from the server
by performing another connection also tried re-producing the exception.
This is probably not necessary as we can get the reason for failure
from the original exception.

Also, produce full chain differently. Before, we would take the whole
chain, and append (leaf is to the left) the root certificate, if any
certificates from the trusted store could verify the rightmost cert.
This works for most cases, but here's a problematic case.

For current websites that get their certs from Let's Encrypt,
Weechat-Android would show the chain as:
  * CN=website.com
  * CN=R3,O=Let's Encrypt
  * CN=ISRG Root X1,O=Internet Security Research Group
  * CN=DST Root CA X3,O=Digital Signature Trust Co. [self-signed]

While the browser would show it as:
  * CN=website.com
  * CN=R3,O=Let's Encrypt
  * CN=ISRG Root X1,O=Internet Security Research Group [self-signed]

Note that the two ISRG certificates are different. Both chains were
valid; on older devices that are not aware of ISRG the CA the first one
was used, and on newer the shorter one is used. This says “were” because
the DST cert expired in September 2001.

Therefore we:
  * Use the full certificate chain produced by the system,
    which is possible if it is valid.
  * If the chain can't be verified, display the chain as is.
    The invalid chain may contain certificates out of order,
    certificates that have nothing to do with the server,
    expired certificates, etc.

See also:
https://letsencrypt.org/docs/dst-root-ca-x3-expiration-september-2021