Obtain certificate chains directly from failing connections
We used to make an additional request to get the server certificate chain.
The idea was to make a request that will ignore any SSL errors,
to have a working SSL session object, from which we can get the chain.

Such a session doesn't follow the regular negotiation,
and so the resulting trusted chain could be incorrect,
as the system might only use a part of the chain.

Also the way we determined whether it was the pin requirement
that failed the connection attempt was not using actual auth type
and hence incorrect.