Rework ping logic
Previously, the code had lines like this:

    SystemClock.elapsedRealtime() - lastMessageReceivedAt > P.pingIdleTime

This logic is fine, but it only works well if broadcast receiver always
receives broadcasts at the exact scheduled time. In practice, sometimes
these can be delayed by several seconds. This is probably not an issue in
most cases, but this can be easily fixed.

Now we record the time of the last alarm and see if we received any messages
since then.