Request permission to schedule exact alarms
Android 12 (API 31) introduced a new permission to schedule exact alarms.
It wasn't a big deal then as it was granted by default for new apps. Still,
it could be removed by the user. More importantly, since Android 14 (34)
apps targeting Android ≥ 13 no longer get this permission by default.

So now when user is trying to connect and we lack this permission, we
explain the rationale for the permission to the user, and request it, also
informing the user that they may forgo the permission by disable pinging.

See https://developer.android.com/about/versions/14/changes/schedule-exact-alarms