Insert spaces around share spans when textifying
When inserting images & other media into the chat input, we add spaces
around it to make sure the text doesn't touch the resulting URL. But if the
user edits the text after, the text might touch the image again. This
sometimes is hard to see, especially if the thumbnail and the text are on
separate lines.

To overcome this, also insert spaces around the URL when textifying.