Record input change after drag-n-drop
In buffer fragments, we are saving input for parallel fragments
(i.e. bubbles) on pause and restoring it on resume. On some systems,
particularly on API 27, it's possible that, when dragging from another app,
the target activity isn't actually resumed, hence this input change may fail
to be recorded. To prevent this, explicitly record the change after loading
the thumbnails.

This includes a long-postponed conversion of thumbnail logic to coroutines.
This facilitates waiting until the thumbnails are actually loaded and
applied to the string.