Fix Play Store publishing
Apparently pyOpenSSL no longer supports PKCS #12. We are using oauth2client,
which is obsolete and hasn't been updated in a while, that uses pyOpenSSL
under the hood.

After briefly looking around at our options, it seems that--if we want to do
things right--we could:
  * Use google-auth & friends. This doesn't support p12 though, so:
    * Somehow download a new key, or
    * Somehow convert the old key to the new format;
  * Use Gradle Play Publisher, it might support p12?

While we explore these options, let's just temporarily pin pyOpenSSL to a
version that supports PKCS #12.

https://github.com/googleapis/google-auth-library-python/issues/288#issuecomment-456671469
https://github.com/Triple-T/gradle-play-publisher