Update Gradle & AGP to version 8
The changes in gradle.properties were made by the migration tool.

This also disables the weaving of Cats, as transform API were disabled
in AGP. Note that somehow the classes are still present, but actually
triggering transformation results in an error.

Cat weaving is fixed in a different way in the next commit.