Implement a framework for encrypting some preferences
The idea is as such: create a custom version of SharedPreferences that
treats preference keys prefixed with `encrypted` differently, saving them
to an instance of EncryptedSharedPreferences.

To actually use this MultiSharedPreferences in a PreferenceFragmentCompat,
we have to employ a little hack of writing its private field, as for some
reason the fragment does not allow customizing it's shared preferences.