Implement a skeleton for testing preference migration
Roboelectric is used to simplify testing with application context and
platform APIs. It provides these things without running on a real device.

We are using JUnit 5. Roboelectric needs JUnit 4; to make this work,
we use org.junit.vintage:junit-vintage-engine.

For in-memory implementation of shared preferences,
we use io.github.ivanshafran:shared-preferences-mock.

Please ignore this:
  testImplementation("androidx.test:runner:1.5.2")
  testImplementation("org.mockito:mockito-core:5.8.0")