Add backup rules
I would prefer to only use the `include` rules, but this is quite hard when
it comes to shared preferences. The problem is that the default shared
preferences file name contains the application id. While it is possible to
use application id placeholder in the manifest, it is not currently possible
in the resource files. It is also apparently not possible to use string
resources, it just does not work. We could also use a Gradle plugin that
gives us placeholder replacements, but Android Stem seems to be only parsing
string resources, which is neat but doesn't fit our use. The only feasible
way here would be duplicating the resources for every build variant.

Rules for API ≤ 30 can be tested with the following Bash script,
which is a modified version of the script suitable for > 30 found here:
at https://developer.android.com/guide/topics/data/testingbackup

    #!/bin/bash -eu
    : "${1?"Usage: $0 package name"}"

    adb shell bmgr enable true

    # $ adb shell bmgr list transports
    #     android/com.android.internal.backup.LocalTransport
    #     com.google.android.gms/.backup.BackupTransportService
    adb shell bmgr transport android/com.android.internal.backup.LocalTransport | grep -q "Selected transport" || (echo "Error: error selecting local transport"; exit 1)

    adb shell settings put secure backup_local_transport_parameters 'is_encrypted=true'
    adb shell bmgr backupnow "$1" | grep -F "Package $1 with result: Success" || (echo "Backup failed"; exit 1)

    apk_path_list=$(adb shell pm path "$1")
    OIFS=$IFS
    IFS=$'\n'
    apk_number=0
    for apk_line in $apk_path_list
    do
        (( ++apk_number ))
        apk_path=${apk_line:8:1000}
        # MSYS_NO_PATHCONV=1 allows running this script in Git Bash
        # Copy before pulling to avoid “remote object ... does not exist” errors
        # See this answer josemigallas https://stackoverflow.com/questions/41150038/#43157127
        MSYS_NO_PATHCONV=1 adb shell cp "$apk_path" "/storage/emulated/0/Download/temp_base.apk"
        MSYS_NO_PATHCONV=1 adb pull "/storage/emulated/0/Download/temp_base.apk" "myapk${apk_number}.apk"
    done
    IFS=$OIFS
    adb shell pm uninstall --user 0 "$1"
    apks=$(seq -f 'myapk%.f.apk' 1 $apk_number)

    adb install -t "$apks"

    adb shell bmgr transport com.google.android.gms/.backup.BackupTransportService
    rm $apks

    echo "Done"