Do not try to emojify using `EmojiCompat` if load failed
We used to try emojifying regardless of EmojiCompat load state. This was not
a big issue, as `EmojiCompat` usually loads fast enough. However, on some
systems loading dynamic fonts is not available, which resulted in errors
printed for every emojification attempt, i. e. for every single line.

This change makes emojifications occur only when `EmojiCompat` is fully
initialized and loaded. Also, initialization status is printed to logcat.

Closes #561