Add off-screen hot buffer indicators in buffer list
Indicators are located at the top and the bottom of the buffer list,
touching window insents, if any. They are shown if there's a hot buffer
above or below the visible buffers, including buffers that are partially
visible. When pressed, they cause the buffer list to scroll to the next hot
buffer in the direction of the button, trying to center the buffer.

The indicators are image buttons with background that is transparent while
the button is not pressed. Both the image and the background are rounded
rectangles. The image is a combination of the badge that we use showing the
number of highlights for a buffer in the buffer list, and the double arrow
that we use for jumping to bottom in the chat.