Add an option to not put hot buffers on top
This changes the boolean preference “Sort buffer list” to a list preference,
with an option to only sort buffers by number.

This does NOT migrate the old setting!

Also, this changes `sortByHotAndMessageCountComparator` to also sort by
number. Not sorting by number could result in buffers appearing out of order
if they were moved. Not a big issue.

Closes #541