In buffer list, only scroll to *new* hot buffers
Before, we would scroll to top whenever we had *any* hot buffers after any
change in buffer list. Now, we only scroll only when:

* Buffer list is not visible
* A buffer becomes hot when it wasn't hot before

The target buffer is centered if possible.

Also, the badge on the bell icon now shows the number of hot buffers, not
the number of new hot messages.