Fix bad behavior when `Lines.replaceLine()` reveals a hidden line
Also add some tests for the method and other around it.

This moves the logic of adding a squiggle line from `addLast()` to
`updateLastLineInfo()`, simplifying things, however, now in some cases the
squiggle lines might become doubled. This also allows, or rather simplifies,
having squiggle lines that are properly `isVisible`, which helps with
preserving them when converting from `unfiltered` to `filtered`. This
conversion becomes wanted as with `replaceLine()` some of the lines may become
visible.

For the same reason, `replaceLine()` also sets skips now.

This also removes `skipFilteredOffset` and `skipUnfilteredOffset` as I just
could not figure out what these were for. Should've added a comment. Might
simply be obsolete logic that stayed due to poor refactoring.

Overall, this is a dangerous change for it touches complex code that has
been working well for quite some time. Ideally it should be rewritten in
some way. This could take a lot of effort and testing.

P.S. The immediate problem that prompted this change was the following crash:

	Exception java.util.NoSuchElementException: ArrayDeque is empty.
	  at kotlin.collections.ArrayDeque.removeFirst (SourceFile:38)
	  at com.ubergeek42.WeechatAndroid.relay.Lines.addLast (SourceFile:124)
	  at com.ubergeek42.WeechatAndroid.relay.BufferList$$ExternalSyntheticLambda0.handleMessage$com$ubergeek42$WeechatAndroid$relay$BufferList$$ExternalSyntheticLambda5 (SourceFile:91)
	  at com.ubergeek42.WeechatAndroid.relay.BufferList$$ExternalSyntheticLambda0.handleMessage (SourceFile:138)
	  at com.ubergeek42.weechat.relay.connection.RelayConnection$$ExternalSyntheticLambda7.run (SourceFile:111)
	  at androidx.core.app.ActivityCompat$$ExternalSyntheticLambda0.run (SourceFile:33)
	  at com.ubergeek42.weechat.relay.connection.Utils$FriendlyThread.run (SourceFile:8)

This happened at the end of the line,

    if (unfiltered.removeFirst().isVisible) filtered.removeFirst()

Here, `filtered` can be empty even if `unfiltered` has a visible line. This
happens if smart filter unhides a join after a message in an otherwise
empty buffer.